refactor unerinfo
